+++ src/Events/PermissionSaving.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Junges\ACL\Events;
+
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Broadcasting\InteractsWithSockets;
+
+class PermissionSaving
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+
+
+    public $permission;
+
+    /**
+     * Create a new event instance.
+     *
+     * @param $attributes
+     */
+    public function __construct($permission)
+    {
+        $this->permission = $permission;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     */
+    public function broadcastOn()
+    {
+        //
+    }
+}